# swagger-spec.json
{"openapi":"3.0.0","info":{"title":"Leslie NestJS API","description":"NestJs API Endpoints","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{},"paths":{"/":{"get":{"operationId":"HelloController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/cats":{"post":{"operationId":"CatsController_create","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"CatsController_forJest","parameters":[],"responses":{"200":{"description":""}}}},"/cats/{id}":{"get":{"operationId":"CatsController_findByIdR","parameters":[{"name":"id","required":true,"in":"path","schema":{}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"CatsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/cats/query-builder/{id}":{"get":{"operationId":"CatsController_findByIdQ","parameters":[{"name":"id","required":true,"in":"path","schema":{}}],"responses":{"200":{"description":""}}}},"/cats/query-builder-fields/{id}":{"get":{"operationId":"CatsController_findbyIdQFields","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/cats/{name}":{"put":{"operationId":"CatsController_update","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/users":{"post":{"operationId":"UsersController_create","parameters":[],"responses":{"201":{"description":""}}}},"/photos":{"post":{"operationId":"PhotosController_create","parameters":[],"responses":{"201":{"description":""}}}},"/heros/{id}/kill":{"post":{"operationId":"HeroesGameController_killDragon","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/heros/{id}/slay":{"post":{"operationId":"HeroesGameController_slayDragon","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/heros/{id}/add-item":{"post":{"operationId":"HeroesGameController_addItem","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/heros":{"post":{"operationId":"HeroesGameController_createHero","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"HeroesGameController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/heros/{id}":{"delete":{"operationId":"HeroesGameController_deleteHero","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"get":{"operationId":"HeroesGameController_findHero","parameters":[],"responses":{"200":{"description":""}}}},"/sample2_post":{"post":{"operationId":"Sample2PostController_create","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"Sample2PostController_findPosts","parameters":[],"responses":{"200":{"description":""}}}},"/sample3_post":{"post":{"operationId":"Sample3ostController_create","parameters":[],"responses":{"201":{"description":""}}}},"/sample3_post/create":{"post":{"operationId":"Sample3ostController_insertNew","parameters":[],"responses":{"201":{"description":""}}}}}}